+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix `ProfileX.sjson` detection on Windows when the `Documents` directory has been moved elsewhere.
+
 ## [v1.4.4] - 2021-10-03
 
 ### Changed