+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+-   Implement bypass for V1.38239 (released on 2021-10-07) breaking monitor resolution detection logic even when unmodded. Hephaistos fixes the bug by bypassing this and forcing in its own custom resolution.
 -   Fix `ProfileX.sjson` detection on Windows when the `Documents` directory has been moved elsewhere.
 
 ## [v1.4.4] - 2021-10-03